add Drink component and some styles for DrinkList.
