+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/index.js":"1","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/containers/CardList.js":"2","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/components/App.js":"3"},{"size":451,"mtime":1611325570773,"results":"4","hashOfConfig":"5"},{"size":624,"mtime":1611329093623,"results":"6","hashOfConfig":"5"},{"size":167,"mtime":1611326998610,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"sxka6j",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/index.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/containers/CardList.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/components/App.js",[]]
+[{"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/index.js":"1","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/components/App.js":"2","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/containers/DrinkList.js":"3","/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/components/Drink.js":"4"},{"size":451,"mtime":1611325570773,"results":"5","hashOfConfig":"6"},{"size":167,"mtime":1611326998610,"results":"7","hashOfConfig":"6"},{"size":814,"mtime":1611500141737,"results":"8","hashOfConfig":"6"},{"size":84,"mtime":1611500326610,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"sxka6j",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":1,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"16","messages":"17","errorCount":4,"warningCount":0,"fixableErrorCount":2,"fixableWarningCount":0,"source":null},"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/index.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/components/App.js",[],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/containers/DrinkList.js",["18"],"/media/praz/0EA40D600EA40D60/microverse projects/REACT-REDUX/react-capstone/src/components/Drink.js",["19","20","21","22"],{"ruleId":"23","severity":2,"message":"24","line":3,"column":19,"nodeType":"25","endLine":3,"endColumn":40},{"ruleId":"26","severity":2,"message":"27","line":1,"column":8,"nodeType":"28","messageId":"29","endLine":1,"endColumn":13},{"ruleId":"26","severity":2,"message":"30","line":3,"column":18,"nodeType":"28","messageId":"29","endLine":3,"endColumn":23},{"ruleId":"31","severity":2,"message":"32","line":5,"column":2,"nodeType":"33","messageId":"34","endLine":6,"endColumn":1,"fix":"35"},{"ruleId":"36","severity":2,"message":"37","line":7,"column":22,"nodeType":"38","messageId":"39","fix":"40"},"import/extensions","Missing file extension for \"../components/Drink\"","Literal","no-unused-vars","'React' is defined but never used.","Identifier","unusedVar","'drink' is defined but never used.","semi","Missing semicolon.","VariableDeclaration","missingSemi",{"range":"41","text":"42"},"eol-last","Newline required at end of file but not found.","Program","missing",{"range":"43","text":"44"},[61,61],";",[84,84],"\n"]
