+++ src/components/Drink.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Drink = ({ drink }) => {
+  
+}
+
+export default Drink;
