+++ src/styles/DrinkList.css
@@ -1,0 +1,4 @@
+.drink-list {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+}
